pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "build",
    "twine",
    "black",
]

[project]
name = "useml"
version = "0.1.0"
description = "Minimalist ML experiment versioning for PyTorch — checkpointing, source archiving, and model isolation."
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "Enzo-py", email = "85191552+Enzo-py@users.noreply.github.com"}]
dependencies = [
    "torch>=2.0.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["useml*"]

[project.urls]
"Homepage" = "https://github.com/Enzo-py/useml"
"Bug Tracker" = "https://github.com/Enzo-py/useml/issues"
